Submit all results instead of trying to guess when they're stale.
Simply forgetting about solutions after a long poll comes in can cause pools to
lose block solutions and miners to lose shares. Not all long polls mean that
all past work is useless - for example, merged mining pools might trigger a
long poll on all of the different chains, but a miner's results might be a
block solution to one of the other unchanged chains.